Dev e

+++ Exercicios/Ep01/floatType.h
@@ -24,6 +24,8 @@ int AlmostEqualRelative(Float_t A, Float_t B);
 
 int AlmostEqualUlps(Float_t A, Float_t B, int maxULPs);
 
-double erro_abs(double a, double b);
+//Função que calcula Erro Absoluto
+float erro_abs(float a, float b);
 
-double erro_rel(double a , double b);
+//Função que Calcula Erro Relativo
+float erro_rel(float a , float b);
